infra/requirements.txt
jsii
cattrs

##
## CDK Modules
##
aws-cdk.core
aws-cdk.aws_lambda
aws-cdk.aws_ecr
aws-cdk.aws_apigateway